feat: 회원 여부 확인 메서드 구현
- 깃허브 ID 타입을 Long으로 변경

+++ src/main/kotlin/com/ixfp/gitmon/db/member/MemberRepository.kt
@@ -6,7 +6,7 @@ import java.util.Optional
 interface MemberRepository : JpaRepository<MemberEntity, Long> {
     override fun findById(memberId: Long): Optional<MemberEntity>
 
-    fun findByGithubId(githubId: Int): MemberEntity?
+    fun findByGithubId(githubId: Long): MemberEntity?
 
     fun findByGithubUsername(githubUsername: String): List<MemberEntity>
 }
